server: Make apc_param_t always 64-bit.

+++ server/async.c
@@ -266,8 +266,8 @@ void async_set_result( struct object *obj, unsigned int status, unsigned int tot
             memset( &data, 0, sizeof(data) );
             data.type         = APC_USER;
             data.user.func    = async->data.apc;
-            data.user.args[0] = (apc_param_t)async->data.arg;
-            data.user.args[1] = (apc_param_t)async->data.iosb;
+            data.user.args[0] = (apc_param_t)(unsigned long)async->data.arg;
+            data.user.args[1] = (apc_param_t)(unsigned long)async->data.iosb;
             data.user.args[2] = 0;
             thread_queue_apc( async->thread, NULL, &data );
         }
